Revert "- Resolved issues #6, #7, #9, #10"
This reverts commit 47520a3deaa022b053a62be09734f3dec79f8bc1.

+++ Models/OrderDetail.cs
@@ -3,11 +3,12 @@ namespace quanlybanhang_nmcnpm.Models;
 public class OrderDetail
 {
     public int SoLuong { get; set; }
-    public decimal DonGiaBan { get; set; }
-    public decimal ThanhTien { get; set; }
+    public int DonGiaBan { get; set; }
+    public int ThanhTien { get; set; }
     public int MaDonHang { get; set; }
     public int MaHang { get; set; }
 
+    // Navigation Properties
     public virtual Order? Order { get; set; }
     public virtual Product? Product { get; set; }
 }
